+++ CHANGES.md
@@ -5,3 +5,4 @@
 
 * Add support for JMS-backed `Stapfen::Worker` classes
 * Deep copy the configuration passed into `Stomp::Client` to work-around [stomp #80](https://github.com/stompgem/stomp/issues/80)
+* Support per-instance log configuration [#3](https://github.com/lookout/stapfen/issues/3)